+++ src/main/java/com/codeborne/selenide/impl/FileHelper.java
@@ -3,12 +3,15 @@ package com.codeborne.selenide.impl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+
+import static org.apache.commons.io.FileUtils.cleanDirectory;
 
 @ParametersAreNonnullByDefault
 public final class FileHelper {
@@ -24,7 +27,7 @@ public final class FileHelper {
   }
 
   public static void copyFile(InputStream in, File targetFile) throws IOException {
-    ensureFolderExists(targetFile);
+    ensureParentFolderExists(targetFile);
 
     try (FileOutputStream out = new FileOutputStream(targetFile)) {
       byte[] buffer = new byte[1024];
@@ -35,13 +38,24 @@ public final class FileHelper {
     }
   }
 
-  private static void ensureFolderExists(File targetFile) {
-    File folder = targetFile.getParentFile();
+  public static void ensureParentFolderExists(File targetFile) {
+    ensureFolderExists(targetFile.getParentFile());
+  }
+
+  @Nonnull
+  public static File ensureFolderExists(File folder) {
     if (!folder.exists()) {
-      log.info("Creating folder: {}", folder);
+      log.info("Creating folder: {}", folder.getAbsolutePath());
       if (!folder.mkdirs()) {
-        log.error("Failed to create {}", folder);
+        log.error("Failed to create folder: {}", folder.getAbsolutePath());
       }
+    }
+    return folder;
+  }
+
+  public static void cleanupFolder(File folder) throws IOException {
+    if (folder.isDirectory()) {
+      cleanDirectory(folder);
     }
   }
 }