+++ src/test/java/com/codeborne/selenide/DriverStub.java
@@ -5,6 +5,8 @@ import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.interactions.Actions;
 
+import java.io.File;
+
 /**
  * A dummy `Driver` implementation used in tests.
  */
@@ -13,6 +15,7 @@ public class DriverStub implements Driver {
   private final Browser browser;
   private final WebDriver webDriver;
   private final SelenideProxyServer proxy;
+  private final File browserDownloadsFolder = new File("build/downloads/45");
 
   public DriverStub() {
     this("zopera");
@@ -57,6 +60,11 @@ public class DriverStub implements Driver {
   @Override
   public WebDriver getAndCheckWebDriver() {
     return webDriver;
+  }
+
+  @Override
+  public File browserDownloadsFolder() {
+    return browserDownloadsFolder;
   }
 
   @Override