+++ src/test/java/com/codeborne/selenide/drivercommands/WebDriverWrapperTest.java
@@ -4,6 +4,8 @@ import com.codeborne.selenide.SelenideConfig;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebDriver;
 
+import java.io.File;
+
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -13,7 +15,7 @@ class WebDriverWrapperTest {
   @Test
   void close_closesTheBrowser() {
     WebDriver webDriver = mock(WebDriver.class);
-    WebDriverWrapper driver = new WebDriverWrapper(new SelenideConfig(), webDriver, null);
+    WebDriverWrapper driver = new WebDriverWrapper(new SelenideConfig(), webDriver, null, new File("build/downloads/135"));
 
     driver.close();