+++ src/test/java/com/codeborne/selenide/webdriver/OperaDriverFactoryTest.java
@@ -9,19 +9,21 @@ import org.openqa.selenium.InvalidArgumentException;
 import org.openqa.selenium.Proxy;
 import org.openqa.selenium.opera.OperaOptions;
 
+import java.io.File;
 import java.util.Map;
 
 import static org.mockito.Mockito.mock;
 
 class OperaDriverFactoryTest implements WithAssertions {
   private final Proxy proxy = mock(Proxy.class);
+  private final File browserDownloadsFolder = new File("build/downlao");
   private final SelenideConfig config = new SelenideConfig().headless(false);
   private final Browser browser = new Browser(config.browser(), config.headless());
 
   @Test
   void browserBinaryCanBeSet() {
     config.browserBinary("c:/browser.exe");
-    Capabilities caps = new OperaDriverFactory().createCapabilities(config, browser, proxy);
+    Capabilities caps = new OperaDriverFactory().createCapabilities(config, browser, proxy, browserDownloadsFolder);
     Map options = (Map) caps.asMap().get(OperaOptions.CAPABILITY);
     assertThat(options.get("binary"))
       .isEqualTo("c:/browser.exe");
@@ -30,7 +32,7 @@ class OperaDriverFactoryTest implements WithAssertions {
   @Test
   void headlessCanNotBeSet() {
     config.headless(true);
-    assertThatThrownBy(() -> new OperaDriverFactory().createCapabilities(config, browser, proxy))
+    assertThatThrownBy(() -> new OperaDriverFactory().createCapabilities(config, browser, proxy, browserDownloadsFolder))
       .isInstanceOf(InvalidArgumentException.class);
   }
 }