+++ statics/src/test/java/integration/BrowserLogsTest.java
@@ -10,7 +10,6 @@ import static com.codeborne.selenide.Selectors.byText;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.getWebDriverLogs;
 import static com.codeborne.selenide.WebDriverRunner.hasWebDriverStarted;
-import static com.codeborne.selenide.WebDriverRunner.isChrome;
 import static com.codeborne.selenide.WebDriverRunner.isFirefox;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assumptions.assumeFalse;
@@ -19,8 +18,9 @@ import static org.openqa.selenium.logging.LogType.BROWSER;
 class BrowserLogsTest extends IntegrationTest {
   @BeforeEach
   void setUp() {
-    // Firefox says `UnsupportedCommandException: POST /session/b493bc56.../log did not match a known command`
-    assumeFalse(isFirefox());
+    assumeFalse(isFirefox(),
+      "Firefox says `UnsupportedCommandException: POST /session/b493bc56.../log did not match a known command`"
+    );
 
     if (hasWebDriverStarted()) {
       getWebDriverLogs(BROWSER); // clear logs
@@ -30,7 +30,6 @@ class BrowserLogsTest extends IntegrationTest {
 
   @Test
   void canGetWebDriverBrowserConsoleLogEntry() {
-    assumeFalse(isChrome());
     $(byText("Generate JS Error")).click();
     List<String> webDriverLogs = getWebDriverLogs(BROWSER, Level.ALL);